progress

+++ testament/categories.nim
@@ -67,9 +67,10 @@ proc icTests(r: var TResults; testsDir: string, cat: Category, options: string) 
 
   const tempExt = "_temp.nim"
   for it in walkDirRec(testsDir / "ic"):
-    if it.endsWith(tempExt):
-      removeFile(it)
-    elif isTestFile(it):
+    #if it.endsWith(tempExt):
+    #  removeFile(it)
+    #elif isTestFile(it):
+    if isTestFile(it) and not it.endsWith(tempExt):
       let nimcache = nimcacheDir(it, options, getTestSpecTarget())
       removeDir(nimcache)
